TPI-341 вынес RoundedCheckbox, удалил формы
