+++ src/components/form/Checkbox/index.less
@@ -1,2 +1,3 @@
 @import "../../../assets/css/bundle.less";
 @import "cform__checkbox.less";
+@import "cform__checkbox--rounded.less";
