+++ src/components/form/Checkbox/index.ts
@@ -1,4 +1,4 @@
-import {Checkbox} from './Checkbox';
+import {Checkbox} from 'components/form/Checkbox/Checkbox';
 import {CheckboxField} from './CheckboxField';
 
-export {Checkbox, CheckboxField};
+export {CheckboxField, Checkbox};
