--- src/components/forms/Checkbox/index.less
@@ -1,2 +1,0 @@
-@import "../../../assets/css/bundle.less";
-@import "./ccheckbox.less";
