--- src/components/forms/Checkbox/index.ts
@@ -1,2 +1,0 @@
-export {Checkbox} from './Checkbox';
-export type {ICheckboxProps} from './Checkbox';
