--- src/components/forms/Forms/index.ts
@@ -1,4 +1,0 @@
-import {IFormProps, OldForm} from './OldForm';
-
-export {OldForm};
-export type {IFormProps};
