--- src/components/forms/Input/index.ts
@@ -1,4 +1,0 @@
-import {IInputProps, Input} from './Input';
-
-export {Input};
-export type {IInputProps};
