--- src/components/forms/Radio/index.less
@@ -1,2 +1,0 @@
-@import "../../../assets/css/bundle.less";
-@import "./cradio.less";
