--- src/components/forms/Radio/index.ts
@@ -1,2 +1,0 @@
-export {Radio} from './Radio';
-export type {IRadioProps} from './Radio';
