--- src/components/forms/RadioButton/index.less
@@ -1,1 +1,0 @@
-@import "./cradiobutton.less";
