--- src/components/forms/RadioButton/index.ts
@@ -1,1 +1,0 @@
-export {RadioButton} from './RadioButton';
