--- src/components/forms/Textarea/index.ts
@@ -1,3 +1,0 @@
-import {Textarea} from './Textarea';
-
-export {Textarea};
