--- src/components/forms/enums.ts
@@ -1,7 +1,0 @@
-export enum State {
-    disabled = 'disabled',
-    active = 'active',
-    error = 'error',
-    success = 'success',
-    idle = 'idle',
-}
