--- src/components/forms/index.tsx
@@ -1,8 +1,0 @@
-import {Input} from './Input';
-import {Textarea} from './Textarea';
-import {RadioButton} from './RadioButton';
-import {Radio} from './Radio';
-import {Checkbox} from './Checkbox';
-import {OldForm} from './Forms';
-
-export {Input, Textarea, RadioButton, Radio, Checkbox, OldForm};
