+++ src/index.tsx
@@ -8,9 +8,6 @@ export * from './components/modal';
 export * from './components/logo';
 export * from './components/dropdown';
 export * from './components/list';
-export * from './components/forms';
-import * as Forms from './components/forms';
-export {Forms};
 export * from './components/loader';
 export * from './components/banner';
 export * from './components/tooltip';
